Refactor login and installment views and templates
