+++ organizer_py/settings.py
@@ -137,4 +137,4 @@ STATIC_URL = "static/"
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
 LOGIN_REDIRECT_URL = "users:home"
 LOGOUT_REDIRECT_URL = "users:home"
-LOGIN_REDIRECT_URL = "users/login"
+LOGIN_URL = "/users/login"
